cpu, stm32_common: fix sign-compare in periph/timer

+++ cpu/stm32_common/periph/timer.c
@@ -93,7 +93,7 @@ int timer_init(tim_t tim, unsigned long freq, timer_cb_t cb, void *arg)
 
 int timer_set_absolute(tim_t tim, int channel, unsigned int value)
 {
-    if (channel >= TIMER_CHAN) {
+    if (channel >= (int)TIMER_CHAN) {
         return -1;
     }
 
@@ -106,7 +106,7 @@ int timer_set_absolute(tim_t tim, int channel, unsigned int value)
 
 int timer_clear(tim_t tim, int channel)
 {
-    if (channel >= TIMER_CHAN) {
+    if (channel >= (int)TIMER_CHAN) {
         return -1;
     }
